Updated PathOptimize's write primitive points method to use ShowImage's write method.  This involved passing in the vectors in ShowImage's write method (thus the modification in ShowImage).
Stopped the creation of non-location endpoints in the optimized graph.  This change was implemented by using isGraphPoint() in PathPoints.  The change required an adjustment in PointTraversal (which was using isGraphPoint()...this should probably be changed).  It also required a change in the overal conversion that basically involved dumping the newly created edge when the last PP in the edge wasn't a GraphPoint.

Stopped the output of locations that aren't connected to anything...this is easy to toggle.  It is simply part of one line of code in isGraphPoint()